+++ packages/cubejs-server-core/src/core/DriverDependencies.js
@@ -27,6 +27,7 @@ module.exports = {
   questdb: '@cubejs-backend/questdb-driver',
   materialize: '@cubejs-backend/materialize-driver',
   vertica: '@knowitall/vertica-driver',
+  pinot: '@cubejs-backend/pinot-driver',
   // List for JDBC drivers
   'databricks-jdbc': '@cubejs-backend/databricks-jdbc-driver',
 };