+++ rust/cubesql/cubesql/src/compile/engine/df/wrapper.rs
@@ -518,15 +518,20 @@ impl CubeScanWrapperNode {
                         order_expr,
                         alias,
                         distinct,
-                        ungrouped,
+                        push_to_cube,
                     }) = wrapped_select_node
                     {
                         // TODO support joins
-                        let ungrouped_scan_node = if ungrouped {
+                        let ungrouped_scan_node = if push_to_cube {
                             if let LogicalPlan::Extension(Extension { node }) = from.as_ref() {
                                 if let Some(cube_scan_node) =
                                     node.as_any().downcast_ref::<CubeScanNode>()
                                 {
+                                    if cube_scan_node.request.ungrouped != Some(true) {
+                                        return Err(CubeError::internal(format!(
+                                            "Expected ungrouped CubeScan node but found: {cube_scan_node:?}"
+                                        )));
+                                    }
                                     Some(Arc::new(cube_scan_node.clone()))
                                 } else {
                                     return Err(CubeError::internal(format!(