+++ rust/cubesql/cubesql/src/compile/rewrite/rules/wrapper/projection.rs
@@ -8,10 +8,11 @@ use crate::{
         wrapped_select_having_expr_empty_tail, wrapped_select_joins_empty_tail,
         wrapped_select_order_expr_empty_tail, wrapped_select_subqueries_empty_tail,
         wrapped_select_window_expr_empty_tail, wrapper_pullup_replacer, wrapper_pushdown_replacer,
-        ListType, LogicalPlanLanguage, ProjectionAlias, WrappedSelectAlias, WrappedSelectUngrouped,
-        WrappedSelectUngroupedScan, WrapperPullupReplacerUngrouped,
+        ListType, LogicalPlanLanguage, ProjectionAlias, WrappedSelectAlias,
+        WrappedSelectPushToCube, WrappedSelectUngroupedScan, WrapperPullupReplacerPushToCube,
+        WrapperPushdownReplacerPushToCube,
     },
-    var, var_iter,
+    copy_flag, var, var_iter,
 };
 use egg::{Subst, Var};
 
@@ -25,7 +26,7 @@ impl WrapperRules {
                     wrapper_pullup_replacer(
                         "?cube_scan_input",
                         "?alias_to_cube",
-                        "?ungrouped",
+                        "?push_to_cube",
                         "?in_projection",
                         "?cube_members",
                     ),
@@ -40,42 +41,42 @@ impl WrapperRules {
                     wrapper_pushdown_replacer(
                         "?expr",
                         "?alias_to_cube",
-                        "?ungrouped",
+                        "?pushdown_push_to_cube",
                         "WrapperPullupReplacerInProjection:true",
                         "?cube_members",
                     ),
                     wrapper_pullup_replacer(
                         wrapped_select_subqueries_empty_tail(),
                         "?alias_to_cube",
-                        "?ungrouped",
+                        "?push_to_cube",
                         "WrapperPullupReplacerInProjection:true",
                         "?cube_members",
                     ),
                     wrapper_pullup_replacer(
                         wrapped_select_group_expr_empty_tail(),
                         "?alias_to_cube",
-                        "?ungrouped",
+                        "?push_to_cube",
                         "WrapperPullupReplacerInProjection:true",
                         "?cube_members",
                     ),
                     wrapper_pullup_replacer(
                         wrapped_select_aggr_expr_empty_tail(),
                         "?alias_to_cube",
-                        "?ungrouped",
+                        "?push_to_cube",
                         "WrapperPullupReplacerInProjection:true",
                         "?cube_members",
                     ),
                     wrapper_pullup_replacer(
                         wrapped_select_window_expr_empty_tail(),
                         "?alias_to_cube",
-                        "?ungrouped",
+                        "?push_to_cube",
                         "WrapperPullupReplacerInProjection:true",
                         "?cube_members",
                     ),
                     wrapper_pullup_replacer(
                         "?cube_scan_input",
                         "?alias_to_cube",
-                        "?ungrouped",
+                        "?push_to_cube",
                         "WrapperPullupReplacerInProjection:true",
                         "?cube_members",
                     ),
@@ -83,7 +84,7 @@ impl WrapperRules {
                     wrapper_pullup_replacer(
                         wrapped_select_filter_expr_empty_tail(),
                         "?alias_to_cube",
-                        "?ungrouped",
+                        "?push_to_cube",
                         "WrapperPullupReplacerInProjection:true",
                         "?cube_members",
                     ),
@@ -93,13 +94,13 @@ impl WrapperRules {
                     wrapper_pullup_replacer(
                         wrapped_select_order_expr_empty_tail(),
                         "?alias_to_cube",
-                        "?ungrouped",
+                        "?push_to_cube",
                         "WrapperPullupReplacerInProjection:true",
                         "?cube_members",
                     ),
                     "?select_alias",
                     "WrappedSelectDistinct:false",
-                    "?select_ungrouped",
+                    "?select_push_to_cube",
                     "?select_ungrouped_scan",
                 ),
                 "CubeScanWrapperFinalized:false",
@@ -107,9 +108,10 @@ impl WrapperRules {
             self.transform_projection(
                 "?expr",
                 "?projection_alias",
-                "?ungrouped",
+                "?push_to_cube",
+                "?pushdown_push_to_cube",
                 "?select_alias",
-                "?select_ungrouped",
+                "?select_push_to_cube",
                 "?select_ungrouped_scan",
             ),
         )]);
@@ -141,7 +143,7 @@ impl WrapperRules {
                         wrapper_pullup_replacer(
                             "?cube_scan_input",
                             "?alias_to_cube",
-                            "?ungrouped",
+                            "?push_to_cube",
                             "?in_projection",
                             "?cube_members",
                         ),
@@ -159,42 +161,42 @@ impl WrapperRules {
                     wrapper_pushdown_replacer(
                         "?expr",
                         "?alias_to_cube",
-                        "?ungrouped",
+                        "?pushdown_push_to_cube",
                         "WrapperPullupReplacerInProjection:true",
                         "?cube_members",
                     ),
                     wrapper_pushdown_replacer(
                         "?subqueries",
                         "?alias_to_cube",
-                        "?ungrouped",
+                        "?pushdown_push_to_cube",
                         "WrapperPullupReplacerInProjection:true",
                         "?cube_members",
                     ),
                     wrapper_pullup_replacer(
                         wrapped_select_group_expr_empty_tail(),
                         "?alias_to_cube",
-                        "?ungrouped",
+                        "?push_to_cube",
                         "WrapperPullupReplacerInProjection:true",
                         "?cube_members",
                     ),
                     wrapper_pullup_replacer(
                         wrapped_select_aggr_expr_empty_tail(),
                         "?alias_to_cube",
-                        "?ungrouped",
+                        "?push_to_cube",
                         "WrapperPullupReplacerInProjection:true",
                         "?cube_members",
                     ),
                     wrapper_pullup_replacer(
                         wrapped_select_window_expr_empty_tail(),
                         "?alias_to_cube",
-                        "?ungrouped",
+                        "?push_to_cube",
                         "WrapperPullupReplacerInProjection:true",
                         "?cube_members",
                     ),
                     wrapper_pullup_replacer(
                         "?cube_scan_input",
                         "?alias_to_cube",
-                        "?ungrouped",
+                        "?push_to_cube",
                         "WrapperPullupReplacerInProjection:true",
                         "?cube_members",
                     ),
@@ -202,7 +204,7 @@ impl WrapperRules {
                     wrapper_pullup_replacer(
                         wrapped_select_filter_expr_empty_tail(),
                         "?alias_to_cube",
-                        "?ungrouped",
+                        "?push_to_cube",
                         "WrapperPullupReplacerInProjection:true",
                         "?cube_members",
                     ),
@@ -212,13 +214,13 @@ impl WrapperRules {
                     wrapper_pullup_replacer(
                         wrapped_select_order_expr_empty_tail(),
                         "?alias_to_cube",
-                        "?ungrouped",
+                        "?push_to_cube",
                         "WrapperPullupReplacerInProjection:true",
                         "?cube_members",
                     ),
                     "?select_alias",
                     "WrappedSelectDistinct:false",
-                    "?select_ungrouped",
+                    "?select_push_to_cube",
                     "?select_ungrouped_scan",
                 ),
                 "CubeScanWrapperFinalized:false",
@@ -227,9 +229,10 @@ impl WrapperRules {
                 "?alias_to_cube",
                 "?expr",
                 "?projection_alias",
-                "?ungrouped",
+                "?push_to_cube",
+                "?pushdown_push_to_cube",
                 "?select_alias",
-                "?select_ungrouped",
+                "?select_push_to_cube",
                 "?select_ungrouped_scan",
             ),
         )]);
@@ -238,16 +241,18 @@ impl WrapperRules {
         &self,
         expr_var: &'static str,
         projection_alias_var: &'static str,
-        ungrouped_var: &'static str,
+        push_to_cube_var: &'static str,
+        pushdown_push_to_cube_var: &'static str,
         select_alias_var: &'static str,
-        select_ungrouped_var: &'static str,
+        select_push_to_cube_var: &'static str,
         select_ungrouped_scan_var: &'static str,
     ) -> impl Fn(&mut CubeEGraph, &mut Subst) -> bool {
         let expr_var = var!(expr_var);
         let projection_alias_var = var!(projection_alias_var);
-        let ungrouped_var = var!(ungrouped_var);
+        let push_to_cube_var = var!(push_to_cube_var);
+        let pushdown_push_to_cube_var = var!(pushdown_push_to_cube_var);
         let select_alias_var = var!(select_alias_var);
-        let select_ungrouped_var = var!(select_ungrouped_var);
+        let select_push_to_cube_var = var!(select_push_to_cube_var);
         let select_ungrouped_scan_var = var!(select_ungrouped_scan_var);
         move |egraph, subst| {
             Self::transform_projection_impl(
@@ -255,9 +260,10 @@ impl WrapperRules {
                 subst,
                 expr_var,
                 projection_alias_var,
-                ungrouped_var,
+                push_to_cube_var,
+                pushdown_push_to_cube_var,
                 select_alias_var,
-                select_ungrouped_var,
+                select_push_to_cube_var,
                 select_ungrouped_scan_var,
             )
         }
@@ -268,17 +274,19 @@ impl WrapperRules {
         alias_to_cube_var: &'static str,
         expr_var: &'static str,
         projection_alias_var: &'static str,
-        ungrouped_var: &'static str,
+        push_to_cube_var: &'static str,
+        pushdown_push_to_cube_var: &'static str,
         select_alias_var: &'static str,
-        select_ungrouped_var: &'static str,
+        select_push_to_cube_var: &'static str,
         select_ungrouped_scan_var: &'static str,
     ) -> impl Fn(&mut CubeEGraph, &mut Subst) -> bool {
         let alias_to_cube_var = var!(alias_to_cube_var);
         let expr_var = var!(expr_var);
         let projection_alias_var = var!(projection_alias_var);
-        let ungrouped_var = var!(ungrouped_var);
+        let push_to_cube_var = var!(push_to_cube_var);
+        let pushdown_push_to_cube_var = var!(pushdown_push_to_cube_var);
         let select_alias_var = var!(select_alias_var);
-        let select_ungrouped_var = var!(select_ungrouped_var);
+        let select_push_to_cube_var = var!(select_push_to_cube_var);
         let select_ungrouped_scan_var = var!(select_ungrouped_scan_var);
         let meta = self.meta_context.clone();
         move |egraph, subst| {
@@ -293,9 +301,10 @@ impl WrapperRules {
                     subst,
                     expr_var,
                     projection_alias_var,
-                    ungrouped_var,
+                    push_to_cube_var,
+                    pushdown_push_to_cube_var,
                     select_alias_var,
-                    select_ungrouped_var,
+                    select_push_to_cube_var,
                     select_ungrouped_scan_var,
                 )
             } else {
@@ -309,28 +318,43 @@ impl WrapperRules {
         subst: &mut Subst,
         expr_var: Var,
         projection_alias_var: Var,
-        ungrouped_var: Var,
+        push_to_cube_var: Var,
+        pushdown_push_to_cube_var: Var,
         select_alias_var: Var,
-        select_ungrouped_var: Var,
+        select_push_to_cube_var: Var,
         select_ungrouped_scan_var: Var,
     ) -> bool {
         if let Some(_) = &egraph[subst[expr_var]].data.referenced_expr {
+            if !copy_flag!(
+                egraph,
+                subst,
+                push_to_cube_var,
+                WrapperPullupReplacerPushToCube,
+                pushdown_push_to_cube_var,
+                WrapperPushdownReplacerPushToCube
+            ) {
+                return false;
+            }
+
             for projection_alias in
                 var_iter!(egraph[subst[projection_alias_var]], ProjectionAlias).cloned()
             {
-                for ungrouped in
-                    var_iter!(egraph[subst[ungrouped_var]], WrapperPullupReplacerUngrouped).cloned()
+                for push_to_cube in var_iter!(
+                    egraph[subst[push_to_cube_var]],
+                    WrapperPullupReplacerPushToCube
+                )
+                .cloned()
                 {
                     subst.insert(
-                        select_ungrouped_var,
-                        egraph.add(LogicalPlanLanguage::WrappedSelectUngrouped(
-                            WrappedSelectUngrouped(ungrouped),
+                        select_push_to_cube_var,
+                        egraph.add(LogicalPlanLanguage::WrappedSelectPushToCube(
+                            WrappedSelectPushToCube(push_to_cube),
                         )),
                     );
                     subst.insert(
                         select_ungrouped_scan_var,
                         egraph.add(LogicalPlanLanguage::WrappedSelectUngroupedScan(
-                            WrappedSelectUngroupedScan(ungrouped),
+                            WrappedSelectUngroupedScan(push_to_cube),
                         )),
                     );
                     subst.insert(